# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Employee Profile & HRMS System API
  description: |
    Comprehensive HR management and employee profile REST API built on CodeIgniter 4.

    ## Authentication
    All protected endpoints require a valid Bearer JWT token in the `Authorization` header.

    **Flow:**
    1. User logs into HRMS — HRMS issues a signed RS256 JWT
    2. Client sends the HRMS JWT to `POST /auth/sso-login`
    3. API validates the JWT, syncs the user, and returns an `access_token` + `refresh_token`
    4. Client uses the `access_token` for all subsequent calls
    5. When the access token expires, call `POST /auth/refresh`

    ## Roles
    | Role | Scope |
    |------|-------|
    | `admin` | Full access to all modules and all employee data |
    | `hr` | Talent, compliance, reporting — department-scoped |
    | `manager` | Own profile + direct reports + team analytics |
    | `employee` | Own profile only |
    | `system` | Service account for HRMS sync operations |

  version: 1.0.0
  contact:
    name: HR Systems Team
servers:
  - url: http://localhost:8000
    description: Local development (php spark serve)
  - url: http://localhost/employee-profile/backend/public
    description: WAMP local server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - BearerAuth: []

# ============================================================
# Paths
# ============================================================
paths:
...
  /auth/sso-login:
    post:
      tags: [Auth]
      summary: SSO Login
      description: Exchange an HRMS-issued RS256 JWT for an application access token.
      security: []
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: "Bearer <hrms_jwt_token>"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Missing token claims
        '401':
          description: Invalid or missing HRMS token
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: New access token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Refresh token missing
        '401':
          description: Invalid or expired refresh token
  /auth/verify:
    get:
      tags: [Auth]
      summary: Verify access token
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  data:
                    type: object
                    properties:
                      claims:
                        type: object
                      expires_at:
                        type: string
                      expires_in:
                        type: integer
        '401':
          description: Token invalid or expired
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout
      description: Invalidates the current session and logs the event.
      responses:
        '200':
          description: Logged out successfully
        '401':
          description: Unauthorized

  # ----------------------------------------------------------
  # Profile
  # ----------------------------------------------------------
  /profile:
    get:
      tags: [Profile]
      summary: Get own profile
      responses:
        '200':
          description: Full employee profile
        '401':
          description: Unauthorized
    put:
      tags: [Profile]
      summary: Update own profile
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
      responses:
        '200':
          description: Profile updated
        '422':
          description: Validation error
  /profile/{id}:
    get:
      tags: [Profile]
      summary: View another employee's profile (role-scoped)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Employee profile
        '403':
          description: Insufficient permissions
        '404':
          description: Employee not found
  /profile/personal-details:
    get:
      tags: [Profile]
      summary: Get personal details
      responses:
        '200':
          description: Personal details (sensitive fields masked by role)
    put:
      tags: [Profile]
      summary: Update personal details
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated
  /profile/addresses:
    get:
      tags: [Profile]
      summary: List addresses
      responses:
        '200':
          description: Address list
    post:
      tags: [Profile]
      summary: Add address
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                address_type:
                  type: string
                  enum: [Home, Work, Permanent, Other]
                street:
                  type: string
                city:
                  type: string
                state:
                  type: string
                country:
                  type: string
                postal_code:
                  type: string
      responses:
        '201':
          description: Address added
  /profile/addresses/{id}:
    put:
      tags: [Profile]
      summary: Update address
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated
    delete:
      tags: [Profile]
      summary: Delete address
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
  /profile/emergency-contacts:
    get:
      tags: [Profile]
      summary: List emergency contacts
      responses:
        '200':
          description: Contact list
    post:
      tags: [Profile]
      summary: Add emergency contact
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                relationship:
                  type: string
                phone:
                  type: string
                email:
                  type: string
      responses:
        '201':
          description: Contact added
  /profile/emergency-contacts/{id}:
    put:
      tags: [Profile]
      summary: Update emergency contact
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated
  /profile/govt-ids:
    get:
      tags: [Profile]
      summary: List government IDs (masked)
      responses:
        '200':
          description: Govt ID list — sensitive fields masked
    post:
      tags: [Profile]
      summary: Add government ID (stored encrypted)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id_type:
                  type: string
                  enum: [Aadhaar, Passport, PAN, Driving Licence, Other]
                id_number:
                  type: string
                issue_date:
                  type: string
                  format: date
                expiry_date:
                  type: string
                  format: date
      responses:
        '201':
          description: ID added (encrypted at rest)
  /profile/govt-ids/{id}:
    put:
      tags: [Profile]
      summary: Update government ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated
  /profile/bank-details:
    get:
      tags: [Profile]
      summary: List bank details (masked)
      responses:
        '200':
          description: Bank details — account numbers masked
    post:
      tags: [Profile]
      summary: Add bank detail (stored encrypted)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                bank_name:
                  type: string
                account_number:
                  type: string
                ifsc_code:
                  type: string
                account_type:
                  type: string
                  enum: [Savings, Current, Salary]
      responses:
        '201':
          description: Bank detail added
  /profile/bank-details/{id}:
    put:
      tags: [Profile]
      summary: Update bank detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated
...
  /profile/family-dependents:
    get:
      tags: [Profile]
      summary: List family dependents
      responses:
        '200':
          description: Dependents list
    post:
      tags: [Profile]
      summary: Add family dependent
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                relationship:
                  type: string
                date_of_birth:
                  type: string
                  format: date
      responses:
        '201':
          description: Dependent added
  /profile/family-dependents/{id}:
    put:
      tags: [Profile]
      summary: Update family dependent
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated

  # ----------------------------------------------------------
  # Job & Organization
  # ----------------------------------------------------------
...
  /job/information/{id}:
    get:
      tags: [Job]
      summary: Get job info by employee ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Job info
        '403':
          description: Insufficient permissions
  /job/history:
    get:
      tags: [Job]
      summary: Get employment history
      responses:
        '200':
          description: History list
    post:
      tags: [Job]
      summary: Add employment history
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                company_name:
                  type: string
                designation:
                  type: string
                start_date:
                  type: string
                  format: date
                end_date:
                  type: string
                  format: date
      responses:
        '201':
          description: History added
  /job/history/{id}:
    get:
      tags: [Job]
      summary: Get specific employment history record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: History record
...
  /job/org-hierarchy/{id}:
    get:
      tags: [Job]
      summary: Get hierarchy for specific node
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Node hierarchy
...
  /job/promotions:
    get:
      tags: [Job]
      summary: List promotions
      responses:
        '200':
          description: Promotion history
    post:
      tags: [Job]
      summary: Record a promotion
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                from_designation:
                  type: string
                to_designation:
                  type: string
                effective_date:
                  type: string
                  format: date
      responses:
        '201':
          description: Promotion recorded
  /job/transfers:
    get:
      tags: [Job]
      summary: List transfers
      responses:
        '200':
          description: Transfer history
    post:
      tags: [Job]
      summary: Record a transfer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                from_department:
                  type: string
                to_department:
                  type: string
                effective_date:
                  type: string
                  format: date
      responses:
        '201':
          description: Transfer recorded

  # ----------------------------------------------------------
  # Performance
  # ----------------------------------------------------------
  /performance/reviews:
    get:
      tags: [Performance]
      summary: List performance reviews
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: per_page
          schema:
            type: integer
      responses:
        '200':
          description: Review list
    post:
      tags: [Performance]
      summary: Create performance review
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                employee_id:
                  type: integer
                review_period:
                  type: string
                  example: 2025-Annual
                overall_rating:
                  type: number
                  minimum: 1
                  maximum: 5
      responses:
        '201':
          description: Review created
  /performance/reviews/{id}:
    get:
      tags: [Performance]
      summary: Get specific review
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Review details
    put:
      tags: [Performance]
      summary: Update review
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated
...
  /performance/feedback/{id}:
    get:
      tags: [Performance]
      summary: Get specific feedback
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Feedback details
...
  /export/org-chart:
    get:
      tags: [Export]
      summary: Export org chart data as JSON
      parameters:
        - in: query
          name: department
          schema:
            type: string
      responses:
        '200':
          description: Org chart export
...